Rename test_additional_models.py -> test_zoo_models.py
